♻️ Refactor setup route into FSD structure
- Extract business logic to useSetupView hook (123 lines)
- Create SetupView presentation component (159 lines)
- Add SetupPage composition layer (12 lines)
- Refactor routes/setup.tsx to thin route (195→11 lines)
- Add accessibility attributes (aria-label, title)
- Maintain 100% feature parity with original implementation

This completes FSD migration for all user-facing pages (7/7).
