+++ app/pages/setup/ui/SetupPage.tsx
@@ -1,0 +1,12 @@
+import { useSetupView } from '~/widgets/setup-view/model/useSetupView';
+import { SetupView } from '~/widgets/setup-view/ui/SetupView';
+
+/**
+ * Setup page composition layer
+ * Connects business logic hook with presentation component
+ */
+export function SetupPage() {
+  const view = useSetupView();
+
+  return <SetupView {...view} />;
+}
