+++ app/widgets/setup-view/ui/SetupView.tsx
@@ -1,0 +1,159 @@
+import type { FormEvent } from 'react';
+import { Eye, EyeOff, Shield } from 'lucide-react';
+import { Alert, AlertDescription } from '~/components/ui/alert';
+import { Button } from '~/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card';
+import { Input } from '~/components/ui/input';
+import { Label } from '~/components/ui/label';
+
+export interface SetupViewProps {
+  email: string;
+  password: string;
+  confirmPassword: string;
+  showPassword: boolean;
+  showConfirmPassword: boolean;
+  loading: boolean;
+  error: string | null;
+  onEmailChange: (value: string) => void;
+  onPasswordChange: (value: string) => void;
+  onConfirmPasswordChange: (value: string) => void;
+  onTogglePassword: () => void;
+  onToggleConfirmPassword: () => void;
+  onSubmit: (event: FormEvent<HTMLFormElement>) => void;
+}
+
+/**
+ * Pure presentation component for admin account setup
+ * Displays form for creating the first admin account
+ */
+export function SetupView({
+  email,
+  password,
+  confirmPassword,
+  showPassword,
+  showConfirmPassword,
+  loading,
+  error,
+  onEmailChange,
+  onPasswordChange,
+  onConfirmPasswordChange,
+  onTogglePassword,
+  onToggleConfirmPassword,
+  onSubmit,
+}: SetupViewProps) {
+  return (
+    <div className="min-h-screen bg-background flex items-center justify-center p-4">
+      <Card className="w-full max-w-md">
+        <CardHeader className="text-center">
+          <div className="flex justify-center mb-4">
+            <div className="p-3 bg-primary/10 rounded-full">
+              <Shield className="h-8 w-8 text-primary" />
+            </div>
+          </div>
+          <CardTitle className="text-2xl">Welcome to Local Streamer</CardTitle>
+          <CardDescription>
+            Create your admin account to get started
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form onSubmit={onSubmit} className="space-y-4">
+            {error && (
+              <Alert variant="destructive">
+                <AlertDescription>
+                  {error}
+                </AlertDescription>
+              </Alert>
+            )}
+
+            <div className="space-y-2">
+              <Label htmlFor="email">Email Address</Label>
+              <Input
+                id="email"
+                type="email"
+                value={email}
+                onChange={event => onEmailChange(event.target.value)}
+                placeholder="admin@example.com"
+                disabled={loading}
+                autoComplete="email"
+                required
+              />
+            </div>
+
+            <div className="space-y-2">
+              <Label htmlFor="password">Password</Label>
+              <div className="relative">
+                <Input
+                  id="password"
+                  type={showPassword ? 'text' : 'password'}
+                  value={password}
+                  onChange={event => onPasswordChange(event.target.value)}
+                  placeholder="Enter your password"
+                  disabled={loading}
+                  autoComplete="new-password"
+                  required
+                />
+                <Button
+                  type="button"
+                  variant="ghost"
+                  size="sm"
+                  className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                  onClick={onTogglePassword}
+                  disabled={loading}
+                  aria-label={showPassword ? 'Hide password' : 'Show password'}
+                  title={showPassword ? 'Hide password' : 'Show password'}
+                >
+                  {showPassword ? (
+                    <EyeOff className="h-4 w-4" />
+                  ) : (
+                    <Eye className="h-4 w-4" />
+                  )}
+                </Button>
+              </div>
+            </div>
+
+            <div className="space-y-2">
+              <Label htmlFor="confirmPassword">Confirm Password</Label>
+              <div className="relative">
+                <Input
+                  id="confirmPassword"
+                  type={showConfirmPassword ? 'text' : 'password'}
+                  value={confirmPassword}
+                  onChange={event => onConfirmPasswordChange(event.target.value)}
+                  placeholder="Confirm your password"
+                  disabled={loading}
+                  autoComplete="new-password"
+                  required
+                />
+                <Button
+                  type="button"
+                  variant="ghost"
+                  size="sm"
+                  className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                  onClick={onToggleConfirmPassword}
+                  disabled={loading}
+                  aria-label={showConfirmPassword ? 'Hide confirm password' : 'Show confirm password'}
+                  title={showConfirmPassword ? 'Hide confirm password' : 'Show confirm password'}
+                >
+                  {showConfirmPassword ? (
+                    <EyeOff className="h-4 w-4" />
+                  ) : (
+                    <Eye className="h-4 w-4" />
+                  )}
+                </Button>
+              </div>
+            </div>
+
+            <Button type="submit" className="w-full" disabled={loading}>
+              {loading ? 'Creating Account...' : 'Create Admin Account'}
+            </Button>
+          </form>
+
+          <div className="mt-6 text-sm text-muted-foreground text-center">
+            <p>This will create the administrator account for Local Streamer.</p>
+            <p className="mt-1">You can add more users later if needed.</p>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
